Fixed delete type

+++ src/main/java/ru/eltex/project/simpleappointer/controllers/MyRESTController.java
@@ -95,7 +95,7 @@ public class MyRESTController {
         ArrayList<String> req = splitURL.split(object);
         for(String i: req){
             System.out.println(i);
-            dateUtil.deleteData(Long.valueOf(i));
+            dateUtil.deleteData(Integer.valueOf(i));
         }
         return 0;
     }
